+++ PCLParaphernalia/ToolMiscSamples.xaml.cs
@@ -1052,7 +1052,7 @@ namespace PCLParaphernalia
             pdlOptionsStore();
 
             ToolMiscSamplesPersist.saveDataCommon(_indxPDL,
-                                                  (int)_indxSampleType);
+                                                  _indxSampleType);
 
             ToolMiscSamplesPersist.saveDataCommonPDL(
                 "PCL",
@@ -1229,10 +1229,10 @@ namespace PCLParaphernalia
                 _paperMarginsLogicalTop = 0;
 
                 _paperWidthPhysical =
-                    (ushort)(_paperSizeShortEdge);
+                    _paperSizeShortEdge;
 
                 _paperLengthPhysical =
-                    (ushort)(_paperSizeLongEdge);
+                    _paperSizeLongEdge;
 
                 _paperWidthPrintable =
                     (ushort)(_paperSizeShortEdge -
@@ -1247,7 +1247,7 @@ namespace PCLParaphernalia
                              (_paperMarginsLogicalPort * 2));
 
                 _paperLengthLogical =
-                    (ushort)(_paperSizeLongEdge);
+                    _paperSizeLongEdge;
             }
             else
             {
@@ -1255,10 +1255,10 @@ namespace PCLParaphernalia
                 _paperMarginsLogicalTop = 0;
 
                 _paperWidthPhysical =
-                    (ushort)(_paperSizeLongEdge);
+                    _paperSizeLongEdge;
 
                 _paperLengthPhysical =
-                    (ushort)(_paperSizeShortEdge);
+                    _paperSizeShortEdge;
 
                 _paperWidthPrintable =
                     (ushort)(_paperSizeLongEdge -
@@ -1273,7 +1273,7 @@ namespace PCLParaphernalia
                              (_paperMarginsLogicalLand * 2));
 
                 _paperLengthLogical =
-                    (ushort)(_paperSizeShortEdge);
+                    _paperSizeShortEdge;
             }
         }
     }
