+++ PCLParaphernalia/ToolSoftFontGenTTF.cs
@@ -514,7 +514,7 @@ namespace PCLParaphernalia
 
             for (int j = 0; j < sliceSize; j++)
             {
-                uiSub = (byte)Buf[j];
+                uiSub = Buf[j];
                 uiTot = (uiTot << 8) | uiSub;
             }
 
@@ -543,7 +543,7 @@ namespace PCLParaphernalia
 
             for (int j = 0; j < sliceSize; j++)
             {
-                uiSub = (byte)Buf[j];
+                uiSub = Buf[j];
                 uiTot = (uiTot << 8) | uiSub;
             }
 
@@ -572,7 +572,7 @@ namespace PCLParaphernalia
 
             for (int j = 0; j < sliceSize; j++)
             {
-                uiSub = (byte)Buf[j];
+                uiSub = Buf[j];
                 uiTot = (uiTot << 8) | uiSub;
             }
 
@@ -601,7 +601,7 @@ namespace PCLParaphernalia
 
             for (int j = 0; j < sliceSize; j++)
             {
-                uiSub = (byte)Buf[j];
+                uiSub = Buf[j];
                 uiTot = (uiTot << 8) | uiSub;
             }
 
@@ -3115,15 +3115,15 @@ namespace PCLParaphernalia
                                 baseEndCode =
                                     (int)(subTabOffset + 14);
                                 baseStartCode =
-                                    (int)(baseEndCode + fmt4SegCountx2 + 2);
+                                    baseEndCode + fmt4SegCountx2 + 2;
                                 baseIdDelta =
-                                    (int)(baseStartCode + fmt4SegCountx2);
+                                    baseStartCode + fmt4SegCountx2;
                                 baseIdRangeOffset =
-                                    (int)(baseIdDelta + fmt4SegCountx2);
+                                    baseIdDelta + fmt4SegCountx2;
                                 baseGlyphIdArray =
-                                    (int)(baseIdRangeOffset + fmt4SegCountx2);
+                                    baseIdRangeOffset + fmt4SegCountx2;
                                 sizeGlyphIdArray =
-                                    (int)(fmt4Length - x);
+                                    fmt4Length - x;
                             }
                         }
                     }
@@ -4378,7 +4378,7 @@ namespace PCLParaphernalia
                                 " too short for " + nameRecCount +
                                 " encoding records");
                         }
-                        else if ((uint)stringsOffset > tabLength)
+                        else if (stringsOffset > tabLength)
                         {
                             flagOK = false;
