+++ PCLParaphernalia/ToolTrayMap.xaml.cs
@@ -1208,7 +1208,7 @@ namespace PCLParaphernalia
 
             //  trayIdSetStore();
 
-            ToolTrayMapPersist.saveDataCommon((int)_indxPDL);
+            ToolTrayMapPersist.saveDataCommon(_indxPDL);
 
             ToolTrayMapPersist.saveDataPCLOpt(_formAsMacroPCL,
                                                _sheetCtPCL);
